Moving determine_procnum_from_RAM() to rmgpy.rmg.main.
