+++ rmgpy/data/kinetics/family.py
@@ -1072,6 +1072,10 @@ class KineticsFamily(Database):
             logging.info('Must be because you turned off the training depository.')
             return
         
+        # Determine number of parallel processes.
+        from rmgpy.rmg.main import determine_procnum_from_RAM
+        procnum = determine_procnum_from_RAM()
+
         tentries = depository.entries
         
         index = max([e.index for e in self.rules.getEntries()] or [0]) + 1
@@ -1156,10 +1160,6 @@ class KineticsFamily(Database):
             # trainingSet=True used later to does not allow species to match a liquid phase library and get corrected thermo which will affect reverse rate calculation
             item = Reaction(reactants=[Species(molecule=[m.molecule[0].copy(deep=True)], label=m.label) for m in entry.item.reactants],
                              products=[Species(molecule=[m.molecule[0].copy(deep=True)], label=m.label) for m in entry.item.products])
-
-            # Determine number of parallel processes.
-            from rmgpy.rmg.react import determine_procnum_from_RAM
-            procnum = determine_procnum_from_RAM() 
 
             if procnum > 1:
                 # If QMTP and multiprocessing write QMTP files here in parallel.
